Add files via upload
The Central Module now updates the weather and the time every time it starts. Virtual Display no longer collapses because there is nothing in the time file, but sometimes it tries to pull up the time file in the interval of it being deleted and replaced by the time module. Going to fix that.